user_service/requirements.txt
fastapi
uvicorn[standard]
pydantic
passlib[bcrypt]
python-jose[cryptography]
email-validator
pydantic-settings
python-multipart